requirements.txt
# Application
Django==2.1.5
djangorestframework==3.9.0
Pillow==5.4.1
rest_condition

# Authentication
django-allauth
django-oauth-toolkit
django-cors-middleware

# Database
psycopg2-binary

# Code quality
pylint
django-lint
pytest
autopep8